node_modules: Move code to lib/data; update node_modules to reflect only NPM
Normally, the 'node_modules' folder is reserved solely for dependencies
installed from NPM. This is done for a few reasons:
(1) It cleanly separates project third-party dependencies from the project's
    source code
(2) It allows for the 'node_modules' folder to be excluded from development
    tools, i.e., git, eslint, etc.

This patch moves the geoConst.json data file to the 'data' folder, as the JSON
file is pure data. Utility functions in geoutils and sendNotification have been
moved to a generic 'lib' folder. These could be move to other folders in the
future, if the amount of new functionality demands it.

The .gitignore file has been updated to simply ignore all installed
dependencies in node_modules.
