+++ lib/geoutils.js
@@ -1,5 +1,5 @@
 //var config = require('config.json');
-var geoConst = require('geoConst.json');
+var geoConst = require('../data/geoConst.json');
 
 function between(point, floor, ceiling ) {
 	var ret = false;
