+++ lib/sendNotification.js
